test: consolidate procman_test.go into main_test.go
Mirrors the source file consolidation for consistency. At ~270 lines,
the tests are easily scannable in one file.

TestMain stays at the top as test infrastructure, followed by unit
tests, then the integration test.
